refactor: player id int -> long
- Changed Player ID field type to Long
- Updated type references
- Remove id annotation in embeds (why it even was there...)

+++ src/main/java/com/root7325/javabs/dao/PlayerDAO.java
@@ -6,7 +6,7 @@ import com.root7325.javabs.entity.player.Player;
  * @author root7325 on 17.06.2025
  */
 public interface PlayerDAO {
-    Player getPlayer(int id);
-    Player getPlayer(int id, String token);
+    Player getPlayer(long id);
+    Player getPlayer(long id, String token);
     Player createPlayer();
 }
